Add files via upload
Scripts and notebook for motorPAINT analysis. version as of May 2022. Analysis should be done from a Notebook, using the MOtorPAINT_postDoM module. Subsequently individual neurites can be analyzed with the projections_v2 notebook using Rolling_window_projection_along_neurite script. This is however, not finished and might therefore be unclear and contain mistake.